requirements.txt

flask
werkzeug
PyPDF2
python-docx
nltk
scikit-learn
reportlab
gunicorn